embedding-index/requirements.txt
fastapi
uvicorn[standard]
faiss-cpu   # Use 'faiss-gpu' if you want GPU acceleration
pydantic
requests
